# platform/frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
      "DataMode": {
        "description": "Qualification rules for data consumed by a run.",
        "enum": [
          "current_research",
          "strict_historical"
        ],
        "title": "DataMode",
        "type": "string"
      },
      "DeploymentStage": {
        "description": "Whether and how a run may affect an account.",
        "enum": [
          "research",
          "shadow",
          "paper",
          "limited_live"
        ],
        "title": "DeploymentStage",
        "type": "string"
      },
      "Envelope": {
        "properties": {
          "context": {
            "$ref": "#/components/schemas/ResponseContext"
          },
          "data": {
            "title": "Data"
          }
        },
        "required": [
          "data",
          "context"
        ],
        "title": "Envelope",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ResponseContext": {
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "coverage": {
            "additionalProperties": true,
            "title": "Coverage",
            "type": "object"
          },
          "data_mode": {
            "$ref": "#/components/schemas/DataMode"
          },
          "dataset_version_ids": {
            "items": {
              "type": "string"
            },
            "title": "Dataset Version Ids",
            "type": "array"
          },
          "deployment_stage": {
            "$ref": "#/components/schemas/DeploymentStage"
          },
          "model_version_ids": {
            "items": {
              "type": "string"
            },
            "title": "Model Version Ids",
            "type": "array"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "system_as_of": {
            "format": "date-time",
            "title": "System As Of",
            "type": "string"
          },
          "trust_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trust State"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "as_of",
          "system_as_of",
          "data_mode",
          "deployment_stage"
        ],
        "title": "ResponseContext",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Read-only P1 capability and governance API",
    "title": "A-Share Platform Next",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/artifacts": {
      "get": {
        "operationId": "artifacts_api_artifacts_get",
        "parameters": [
          {
            "in": "query",
            "name": "data_mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Data Mode"
            }
          },
          {
            "in": "query",
            "name": "deployment_stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Deployment Stage"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Artifacts"
      }
    },
    "/api/capabilities": {
      "get": {
        "operationId": "capabilities_api_capabilities_get",
        "parameters": [
          {
            "in": "query",
            "name": "data_mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Data Mode"
            }
          },
          {
            "in": "query",
            "name": "deployment_stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Deployment Stage"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Capabilities"
      }
    },
    "/api/datasets": {
      "get": {
        "operationId": "datasets_api_datasets_get",
        "parameters": [
          {
            "in": "query",
            "name": "data_mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Data Mode"
            }
          },
          {
            "in": "query",
            "name": "deployment_stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Deployment Stage"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Datasets"
      }
    },
    "/api/health": {
      "get": {
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/api/identity": {
      "get": {
        "operationId": "identity_api_identity_get",
        "parameters": [
          {
            "in": "query",
            "name": "data_mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Data Mode"
            }
          },
          {
            "in": "query",
            "name": "deployment_stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Deployment Stage"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Identity"
      }
    },
    "/api/runs": {
      "get": {
        "operationId": "runs_api_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "data_mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Data Mode"
            }
          },
          {
            "in": "query",
            "name": "deployment_stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Deployment Stage"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Runs"
      }
    },
    "/api/version": {
      "get": {
        "operationId": "application_version_api_version_get",
        "parameters": [
          {
            "in": "query",
            "name": "data_mode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Data Mode"
            }
          },
          {
            "in": "query",
            "name": "deployment_stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Deployment Stage"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Application Version"
      }
    }
  }
}