# platform/scripts/export_openapi.py
"""Export the deterministic P1 OpenAPI contract for frontend type generation."""

from __future__ import annotations

import json
from pathlib import Path

from a_share_platform.api.app import create_app


def main() -> None:
    directory = Path(__file__).resolve().parents[1] / "frontend" / "src" / "api"
    directory.mkdir(parents=True, exist_ok=True)
    schema = create_app().openapi()
    (directory / "openapi.json").write_text(
        json.dumps(schema, ensure_ascii=False, indent=2, sort_keys=True) + "\n",
        encoding="utf-8",
    )
    get_paths = sorted(
        path
        for path, operations in schema["paths"].items()
        if "get" in operations
    )
    path_contracts = "\n".join(
        f'  "{path}": {{ get: ReadOperation }};'
        for path in get_paths
    )
    (directory / "schema.d.ts").write_text(
        """// Generated by platform/scripts/export_openapi.py. Do not edit manually.
export type DataMode = 'current_research' | 'strict_historical';
export type DeploymentStage = 'research' | 'shadow' | 'paper' | 'limited_live';

export interface ResponseContext {
  as_of: string;
  system_as_of: string;
  data_mode: DataMode;
  deployment_stage: DeploymentStage;
  trust_state: string | null;
  dataset_version_ids: string[];
  model_version_ids: string[];
  run_id: string | null;
  coverage: Record<string, unknown>;
  warnings: string[];
}

export interface Envelope<T = unknown> {
  data: T;
  context: ResponseContext;
}

export interface ProblemDetails {
  type: string;
  title: string;
  status: number;
  detail: string;
  instance: string;
}

export interface ReadOperation {
  responses: { 200: { content: { 'application/json': Envelope } } };
}

export interface paths {
"""
        + path_contracts
        + "\n}\n",
        encoding="utf-8",
    )


if __name__ == "__main__":
    main()